Integrate chart brief generation into gamma_mcp.py
- Moved ChartBriefGenerator functionality from chart_brief_generator.py into gamma_mcp.py
- Added generate_chart_brief() as an MCP tool for creating structured chart briefs
- Implemented ChartType enum with support for 9 chart types (bar, column, line, pie, donut, scatter, bubble, heatmap, area)
- Added chart-specific instructions and formatting for each chart type
- Includes data table formatting and validation
- Follows Chart Brief Template from VICTORIA.md for professional visualizations
- Deleted unused chart_brief_generator.py file
- Chart brief functionality is now accessible as an MCP tool for Victoria
